split into settingsutilities
